+++ lib/web/js/JSUtilsWeb.dart
@@ -3,8 +3,6 @@ library bluetooth_web_js_utils;
 import 'dart:js' as JS;
 import 'dart:js_util' as JSUtil;
 
-import 'package:flutter/foundation.dart';
-
 part 'JSUtilsInterface.dart';
 
 class JSUtils extends JSUtilsInterface {